Moved SW360ClientConfig class to a separate config package.
Issue: eclipse#423

This was done to avoid cyclic dependencies: from the main client
package, there will be references to the SW360XXXClient classes; and
these classes reference the SW360ClientConfig class.
