+++ modules/sw360/sw360-client/src/main/java/org/eclipse/sw360/antenna/sw360/client/auth/SW360AuthenticationClient.java
@@ -16,7 +16,7 @@ package org.eclipse.sw360.antenna.sw360.client.auth;
 import org.eclipse.sw360.antenna.http.api.RequestBuilder;
 import org.eclipse.sw360.antenna.http.api.Response;
 import org.eclipse.sw360.antenna.http.utils.HttpUtils;
-import org.eclipse.sw360.antenna.sw360.client.SW360ClientConfig;
+import org.eclipse.sw360.antenna.sw360.client.config.SW360ClientConfig;
 import org.eclipse.sw360.antenna.sw360.rest.resource.SW360Attributes;
 
 import java.io.IOException;
