+++ modules/sw360/sw360-client/src/main/java/org/eclipse/sw360/antenna/sw360/client/config/SW360ClientConfig.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.sw360.antenna.sw360.client;
+package org.eclipse.sw360.antenna.sw360.client.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.StringUtils;
